+++ src/main/java/de/samply/frontend/dto/Results.java
@@ -13,6 +13,7 @@ public record Results(
         String lastName,
         String url,
         UserProjectState creatorState,
-        ProjectBridgeheadState bridgeheadState
+        ProjectBridgeheadState bridgeheadAdminState,
+        UserProjectState finalUserState
 ) {
 }