+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2025-01-09]
+## [0.0.1 - 2025-01-10]
 ### Added
 - First version of the project
 - Spring Application
@@ -187,3 +187,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Sort Mailing Black List by first and last name
 - Results
 - DRN in exporter label
+- Final User State in results